LayoutTests/wasm/window-postmessage-expected.txt

Test that expected Wasm Module can be sent over window.postMessage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS () => value is 7002
PASS successfullyParsed is true

TEST COMPLETE
